# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a microservice for user management.",
        "title": "Keeper API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/impersonations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Requires impersonation.read (falcon-resolved). Lists active impersonation sessions for audit.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "List active impersonation sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max results (default 50, max 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Result offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/impersonation.ImpersonationSession"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Requires impersonation.create (falcon-resolved). Mints a one-time handoff code for logging in as another user on a registered downstream service. The code is exchanged (cross-origin) for the actual token. Refuses to target a sysadmin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Start an impersonation session",
                "parameters": [
                    {
                        "description": "Target user and audience",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/impersonation.StartImpersonationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/impersonation.StartImpersonationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/active/{session_id}": {
            "get": {
                "description": "Public (no auth) and rate-limited. Returns a boolean only — no identity or tenant data — so downstream services can cheaply enforce revocation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Check whether an impersonation session is active",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opaque session id",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/impersonation.SessionStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/exchange": {
            "post": {
                "description": "Public (no auth) and rate-limited. Redeems a one-time code for a short-lived, audience-scoped impersonation token plus the impersonated user object. Called cross-origin by the target service UI's exchange page.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Exchange a handoff code for an impersonation token",
                "parameters": [
                    {
                        "description": "Handoff code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/impersonation.ExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/impersonation.ExchangeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/logout": {
            "post": {
                "description": "Public (no auth) and rate-limited. Revokes a session by its opaque id — used by the impersonation tab on exit. Knowing the unguessable session id is the capability; revocation only reduces access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Revoke an impersonation session (self-service)",
                "parameters": [
                    {
                        "description": "Session id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/impersonation.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/services": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Requires impersonation.read (falcon-resolved). Returns the registered services a caller can impersonate a user into, for the UI service picker.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "List registered impersonation target services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/impersonation.ServiceInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Requires impersonation.read (falcon-resolved).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Get an impersonation session by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/impersonation.ImpersonationSession"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
        "/impersonations/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Requires impersonation.create (falcon-resolved). Revokes a session server-side; downstream services reject its token on next request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "impersonation"
                ],
                "summary": "Revoke an impersonation session by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/impersonation.ImpersonationSession"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/render.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "impersonation.ExchangeRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "impersonation.ExchangeResponse": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/impersonation.TargetUser"
                }
            }
        },
        "impersonation.ImpersonationSession": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "integer"
                },
                "audience": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "division_id": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "impersonator_user_id": {
                    "type": "integer"
                },
                "read_only": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "target_user_id": {
                    "type": "integer"
                }
            }
        },
        "impersonation.LogoutRequest": {
            "type": "object",
            "required": [
                "session_id"
            ],
            "properties": {
                "session_id": {
                    "type": "string"
                }
            }
        },
        "impersonation.ServiceInfo": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "ui_exchange_url": {
                    "type": "string"
                }
            }
        },
        "impersonation.SessionStatusResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                }
            }
        },
        "impersonation.StartImpersonationRequest": {
            "type": "object",
            "required": [
                "audience",
                "target_user_id"
            ],
            "properties": {
                "audience": {
                    "type": "string"
                },
                "read_only": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "target_user_id": {
                    "type": "integer"
                }
            }
        },
        "impersonation.StartImpersonationResponse": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
...
        "render.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}